requirements.txt
gradio>=4.0.0
pandas
vanna
chromadb
python-dotenv 